requirements.txt
pandas
pytest
streamlit
tabulate
google-genai
